PoC to decouple abs implementation to separate source file
