+++ dpnp/backend/extensions/vm/abs.hpp
@@ -25,72 +25,12 @@
 
 #pragma once
 
-#include <CL/sycl.hpp>
+#include <pybind11/pybind11.h>
+#include <pybind11/stl.h>
 
-#include "common.hpp"
-#include "types_matrix.hpp"
+namespace py = pybind11;
 
-// dpctl tensor headers
-#include "utils/type_dispatch.hpp"
-
-namespace td_ns = dpctl::tensor::type_dispatch;
-
-namespace dpnp::backend::ext::vm
+namespace dpnp::extensions::vm
 {
-template <typename T>
-sycl::event abs_contig_impl(sycl::queue &exec_q,
-                            std::size_t in_n,
-                            const char *in_a,
-                            char *out_y,
-                            const std::vector<sycl::event> &depends)
-{
-    type_utils::validate_type_for_device<T>(exec_q);
-
-    std::int64_t n = static_cast<std::int64_t>(in_n);
-    const T *a = reinterpret_cast<const T *>(in_a);
-    using resTy = typename types::AbsOutputType<T>::value_type;
-    resTy *y = reinterpret_cast<resTy *>(out_y);
-
-    return mkl_vm::abs(exec_q,
-                       n, // number of elements to be calculated
-                       a, // pointer `a` containing input vector of size n
-                       y, // pointer `y` to the output vector of size n
-                       depends);
-}
-
-template <typename fnT, typename T>
-struct AbsContigFactory
-{
-    fnT get()
-    {
-        if constexpr (std::is_same_v<
-                          typename types::AbsOutputType<T>::value_type, void>)
-        {
-            return nullptr;
-        }
-        else {
-            return abs_contig_impl<T>;
-        }
-    }
-};
-
-template <typename fnT, typename T>
-struct AbsStridedFactory
-{
-    fnT get()
-    {
-        return nullptr;
-    }
-};
-
-template <typename fnT, typename T>
-struct AbsTypeMapFactory
-{
-    /*! @brief get typeid for output type of abs(T x) */
-    std::enable_if_t<std::is_same<fnT, int>::value, int> get()
-    {
-        using rT = typename types::AbsOutputType<T>::value_type;
-        return td_ns::GetTypeid<rT>{}.get();
-    }
-};
-} // namespace dpnp::backend::ext::vm
+void init_abs(py::module_ m);
+} // namespace dpnp::extensions::vm
