+++ dpnp/backend/extensions/vm/types_matrix.hpp
@@ -45,23 +45,6 @@ namespace types
 {
 /**
  * @brief A factory to define pairs of supported types for which
- * MKL VM library provides support in oneapi::mkl::vm::abs<T> function.
- *
- * @tparam T Type of input vector `a` and of result vector `y`.
- */
-template <typename T>
-struct AbsOutputType
-{
-    using value_type = typename std::disjunction<
-        dpctl_td_ns::TypeMapResultEntry<T, std::complex<double>, double>,
-        dpctl_td_ns::TypeMapResultEntry<T, std::complex<float>, float>,
-        dpctl_td_ns::TypeMapResultEntry<T, double>,
-        dpctl_td_ns::TypeMapResultEntry<T, float>,
-        dpctl_td_ns::DefaultResultEntry<void>>::result_type;
-};
-
-/**
- * @brief A factory to define pairs of supported types for which
  * MKL VM library provides support in oneapi::mkl::vm::acos<T> function.
  *
  * @tparam T Type of input vector `a` and of result vector `y`.
